Added Normalize
Intended as a stopgap measure for multi-browser support. This should ease the headache of supporting it ourselves for the time being, until we can come up with an in-house solution.

As this does *a lot* of changes, keep an eye out for new and exciting bugs we just were not expecting.