Experimental debug commands
Chrome developer tools just does not like adding new functions, which makes testing code a bit of a hassle for myself and others. These commands hopefully will allow use to just hijack the blank functions for testing code on. 

There is probably a better way, but this is my way until I know of an alternative.